+++ Craftsman/Builders/Tests/Utilities/HttpClientExtensionsBuilder.cs
@@ -24,7 +24,7 @@ public class HttpClientExtensionsBuilder
         return @$"namespace {classPath.ClassNamespace};
 
 using Microsoft.AspNetCore.JsonPatch;
-using System.Text.Json;
+using Newtonsoft.Json;
 using System.Dynamic;
 using System.Net;
 using System.Net.Http.Json;
@@ -66,7 +66,7 @@ public static class HttpClientExtensions
     public static async Task<HttpResponseMessage> PatchJsonRequestAsync<TModel>(this HttpClient client, string url, JsonPatchDocument<TModel> patchDoc)
         where TModel : class
     {{
-        var serializedRecipeToUpdate = JsonSerializer.Serialize(patchDoc);
+        var serializedRecipeToUpdate = JsonConvert.SerializeObject(patchDoc);
 
         var patchRequest = new HttpRequestMessage(new HttpMethod(""PATCH""), url)
         {{