+++ CHANGELOG.md
@@ -86,6 +86,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 * Automatically add a `SuperAdmin` role claim to machines
 
+* Updated services to use Newtonsoft to keep support for patchdocs. `System.Text.Json` is still available if patchdocs aren't needed for a project.
+
 * Added `client_role` to `UserPolicyHandler` role check to accomodate machines with a new scaffolded test.
 
   * **TODO DOCS**. For example with Duende, you can add something like this to get a role. Ideally configed in the DB.